rename test packages to querytests and integrationtests
