+++ x/crosschain/client/integrationtests/inbound_voter_test.go
@@ -1,7 +1,7 @@
 //go:build TESTNET
 // +build TESTNET
 
-package cli_integration_tests
+package integrationtests
 
 import (
 	clitestutil "github.com/cosmos/cosmos-sdk/testutil/cli"
