+++ x/crosschain/client/integrationtests/suite.go
@@ -1,7 +1,7 @@
 //go:build TESTNET
 // +build TESTNET
 
-package cli_integration_tests
+package integrationtests
 
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
