+++ x/crosschain/client/querytests/gas_price.go
@@ -1,75 +1,65 @@
-package cli_query_tests
+package querytests
 
 import (
 	"fmt"
 	"github.com/cosmos/cosmos-sdk/client/flags"
 	clitestutil "github.com/cosmos/cosmos-sdk/testutil/cli"
-	"google.golang.org/grpc/codes"
-
 	tmcli "github.com/tendermint/tendermint/libs/cli"
-	"github.com/zeta-chain/zetacore/testutil/nullify"
 	"github.com/zeta-chain/zetacore/x/crosschain/client/cli"
 	"github.com/zeta-chain/zetacore/x/crosschain/types"
+	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
-	"strconv"
 )
 
-func (s *CliTestSuite) TestShowInTxHashToCctx() {
+func (s *CliTestSuite) TestShowGasPrice() {
 	ctx := s.network.Validators[0].ClientCtx
-	objs := s.state.InTxHashToCctxList
+	objs := s.state.GasPriceList
 	common := []string{
 		fmt.Sprintf("--%s=json", tmcli.OutputFlag),
 	}
 	for _, tc := range []struct {
-		desc       string
-		idInTxHash string
-
+		desc string
+		id   string
 		args []string
 		err  error
-		obj  types.InTxHashToCctx
+		obj  *types.GasPrice
 	}{
 		{
-			desc:       "found",
-			idInTxHash: objs[0].InTxHash,
-
+			desc: "found",
+			id:   objs[0].Index,
 			args: common,
 			obj:  objs[0],
 		},
 		{
-			desc:       "not found",
-			idInTxHash: strconv.Itoa(100000),
-
+			desc: "not found",
+			id:   "not_found",
 			args: common,
-			err:  status.Error(codes.NotFound, "not found"),
+			err:  status.Error(codes.InvalidArgument, "not found"),
 		},
 	} {
+		tc := tc
 		s.Run(tc.desc, func() {
-			args := []string{
-				tc.idInTxHash,
-			}
+			args := []string{tc.id}
 			args = append(args, tc.args...)
-			out, err := clitestutil.ExecTestCLICmd(ctx, cli.CmdShowInTxHashToCctx(), args)
+			out, err := clitestutil.ExecTestCLICmd(ctx, cli.CmdShowGasPrice(), args)
 			if tc.err != nil {
 				stat, ok := status.FromError(tc.err)
 				s.Require().True(ok)
 				s.Require().ErrorIs(stat.Err(), tc.err)
 			} else {
 				s.Require().NoError(err)
-				var resp types.QueryGetInTxHashToCctxResponse
+				var resp types.QueryGetGasPriceResponse
 				s.Require().NoError(s.network.Config.Codec.UnmarshalJSON(out.Bytes(), &resp))
-				s.Require().NotNil(resp.InTxHashToCctx)
-				s.Require().Equal(nullify.Fill(&tc.obj),
-					nullify.Fill(&resp.InTxHashToCctx),
-				)
+				s.Require().NotNil(resp.GasPrice)
+				s.Require().Equal(tc.obj, resp.GasPrice)
 			}
 		})
 	}
 }
 
-func (s *CliTestSuite) TestListInTxHashToCctx() {
+func (s *CliTestSuite) TestListGasPrice() {
 	ctx := s.network.Validators[0].ClientCtx
-	objs := s.state.InTxHashToCctxList
-	cctxCount := len(s.state.CrossChainTxs)
+	objs := s.state.GasPriceList
 	request := func(next []byte, offset, limit uint64, total bool) []string {
 		args := []string{
 			fmt.Sprintf("--%s=json", tmcli.OutputFlag),
@@ -89,14 +79,13 @@ func (s *CliTestSuite) TestListInTxHashToCctx() {
 		step := 2
 		for i := 0; i < len(objs); i += step {
 			args := request(nil, uint64(i), uint64(step), false)
-			out, err := clitestutil.ExecTestCLICmd(ctx, cli.CmdListInTxHashToCctx(), args)
+			out, err := clitestutil.ExecTestCLICmd(ctx, cli.CmdListGasPrice(), args)
 			s.Require().NoError(err)
-			var resp types.QueryAllInTxHashToCctxResponse
+			var resp types.QueryAllGasPriceResponse
 			s.Require().NoError(s.network.Config.Codec.UnmarshalJSON(out.Bytes(), &resp))
-			s.Require().LessOrEqual(len(resp.InTxHashToCctx), step)
-			s.Require().Subset(nullify.Fill(objs),
-				nullify.Fill(resp.InTxHashToCctx),
-			)
+			for j := i; j < len(objs) && j < i+step; j++ {
+				s.Assert().Equal(objs[j], resp.GasPrice[j-i])
+			}
 		}
 	})
 	s.Run("ByKey", func() {
@@ -104,29 +93,24 @@ func (s *CliTestSuite) TestListInTxHashToCctx() {
 		var next []byte
 		for i := 0; i < len(objs); i += step {
 			args := request(next, 0, uint64(step), false)
-			out, err := clitestutil.ExecTestCLICmd(ctx, cli.CmdListInTxHashToCctx(), args)
+			out, err := clitestutil.ExecTestCLICmd(ctx, cli.CmdListGasPrice(), args)
 			s.Require().NoError(err)
-			var resp types.QueryAllInTxHashToCctxResponse
+			var resp types.QueryAllGasPriceResponse
 			s.Require().NoError(s.network.Config.Codec.UnmarshalJSON(out.Bytes(), &resp))
-			s.Require().LessOrEqual(len(resp.InTxHashToCctx), step)
-			s.Require().Subset(nullify.Fill(objs),
-				nullify.Fill(resp.InTxHashToCctx),
-			)
+			for j := i; j < len(objs) && j < i+step; j++ {
+				s.Assert().Equal(objs[j], resp.GasPrice[j-i])
+			}
 			next = resp.Pagination.NextKey
 		}
 	})
 	s.Run("Total", func() {
 		args := request(nil, 0, uint64(len(objs)), true)
-		out, err := clitestutil.ExecTestCLICmd(ctx, cli.CmdListInTxHashToCctx(), args)
+		out, err := clitestutil.ExecTestCLICmd(ctx, cli.CmdListGasPrice(), args)
 		s.Require().NoError(err)
-		var resp types.QueryAllInTxHashToCctxResponse
-		fmt.Println(out.String())
+		var resp types.QueryAllGasPriceResponse
 		s.Require().NoError(s.network.Config.Codec.UnmarshalJSON(out.Bytes(), &resp))
 		s.Require().NoError(err)
-		// saving CCTX also adds a new mapping
-		s.Require().Equal(len(objs)+cctxCount, int(resp.Pagination.Total))
-		s.Require().ElementsMatch(nullify.Fill(objs),
-			nullify.Fill(resp.InTxHashToCctx),
-		)
+		s.Require().Equal(len(objs), int(resp.Pagination.Total))
+		s.Require().Equal(objs, resp.GasPrice)
 	})
 }
