+++ x/crosschain/client/querytests/last_block_height.go
@@ -1,4 +1,4 @@
-package cli_query_tests
+package querytests
 
 import (
 	"fmt"
