+++ x/crosschain/client/querytests/permission_flags.go
@@ -1,17 +1,18 @@
-package cli_query_tests
+package querytests
 
 import (
 	"fmt"
 	clitestutil "github.com/cosmos/cosmos-sdk/testutil/cli"
 	tmcli "github.com/tendermint/tendermint/libs/cli"
+	"github.com/zeta-chain/zetacore/testutil/nullify"
 	"github.com/zeta-chain/zetacore/x/crosschain/client/cli"
 	"github.com/zeta-chain/zetacore/x/crosschain/types"
 	"google.golang.org/grpc/status"
 )
 
-func (s *CliTestSuite) TestShowTSS() {
+func (s *CliTestSuite) TestShowPermissionFlags() {
 	ctx := s.network.Validators[0].ClientCtx
-	obj := s.state.Tss
+	obj := s.state.PermissionFlags
 	common := []string{
 		fmt.Sprintf("--%s=json", tmcli.OutputFlag),
 	}
@@ -19,30 +20,30 @@ func (s *CliTestSuite) TestShowTSS() {
 		desc string
 		args []string
 		err  error
-		obj  *types.TSS
+		obj  *types.PermissionFlags
 	}{
 		{
 			desc: "get",
 			args: common,
 			obj:  obj,
-			err:  nil,
 		},
 	} {
-		tc := tc
 		s.Run(tc.desc, func() {
 			var args []string
 			args = append(args, tc.args...)
-			out, err := clitestutil.ExecTestCLICmd(ctx, cli.CmdShowTSS(), args)
+			out, err := clitestutil.ExecTestCLICmd(ctx, cli.CmdShowPermissionFlags(), args)
 			if tc.err != nil {
 				stat, ok := status.FromError(tc.err)
 				s.Require().True(ok)
 				s.Require().ErrorIs(stat.Err(), tc.err)
 			} else {
 				s.Require().NoError(err)
-				var resp types.QueryGetTSSResponse
+				var resp types.QueryGetPermissionFlagsResponse
 				s.Require().NoError(s.network.Config.Codec.UnmarshalJSON(out.Bytes(), &resp))
-				s.Require().NotNil(resp.TSS)
-				s.Require().Equal(tc.obj, resp.TSS)
+				s.Require().NotNil(resp.PermissionFlags)
+				s.Require().Equal(nullify.Fill(&tc.obj),
+					nullify.Fill(&resp.PermissionFlags),
+				)
 			}
 		})
 	}
