+++ x/crosschain/client/querytests/tss.go
@@ -1,4 +1,4 @@
-package cli_query_tests
+package querytests
 
 import (
 	"fmt"
@@ -6,51 +6,43 @@ import (
 	tmcli "github.com/tendermint/tendermint/libs/cli"
 	"github.com/zeta-chain/zetacore/x/crosschain/client/cli"
 	"github.com/zeta-chain/zetacore/x/crosschain/types"
-	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 )
 
-func (s *CliTestSuite) TestShowNodeAccount() {
+func (s *CliTestSuite) TestShowTSS() {
 	ctx := s.network.Validators[0].ClientCtx
-	objs := s.state.NodeAccountList
+	obj := s.state.Tss
 	common := []string{
 		fmt.Sprintf("--%s=json", tmcli.OutputFlag),
 	}
 	for _, tc := range []struct {
 		desc string
-		id   string
 		args []string
 		err  error
-		obj  *types.NodeAccount
+		obj  *types.TSS
 	}{
 		{
-			desc: "found",
-			id:   objs[0].Operator,
+			desc: "get",
 			args: common,
-			obj:  objs[0],
-		},
-		{
-			desc: "not found",
-			id:   "not_found",
-			args: common,
-			err:  status.Error(codes.InvalidArgument, "not found"),
+			obj:  obj,
+			err:  nil,
 		},
 	} {
 		tc := tc
 		s.Run(tc.desc, func() {
-			args := []string{tc.id}
+			var args []string
 			args = append(args, tc.args...)
-			out, err := clitestutil.ExecTestCLICmd(ctx, cli.CmdShowNodeAccount(), args)
+			out, err := clitestutil.ExecTestCLICmd(ctx, cli.CmdShowTSS(), args)
 			if tc.err != nil {
 				stat, ok := status.FromError(tc.err)
 				s.Require().True(ok)
 				s.Require().ErrorIs(stat.Err(), tc.err)
 			} else {
 				s.Require().NoError(err)
-				var resp types.QueryGetNodeAccountResponse
+				var resp types.QueryGetTSSResponse
 				s.Require().NoError(s.network.Config.Codec.UnmarshalJSON(out.Bytes(), &resp))
-				s.Require().NotNil(resp.NodeAccount)
-				s.Require().Equal(tc.obj, resp.NodeAccount)
+				s.Require().NotNil(resp.TSS)
+				s.Require().Equal(tc.obj, resp.TSS)
 			}
 		})
 	}
